klipper-sdk/pyproject.toml
[tool.poetry]
name = "klipper-sdk"
version = "0.1.0"
description = "A Python SDK for the KDE Klipper Clipboard Service (OCS Integrated)"
authors = ["OCS Builder <builder@ocs.local>"]
readme = "README.md"
packages = [{include = "klipper_sdk", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
dbus-python = "^1.3.2"
pydantic = "^2.0"
rich = "^13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21"
ruff = "^0.0.280"
mypy = "^1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
